requirements.txt
cycler==0.11.0
fonttools==4.34.4
joblib==1.1.0
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.23.2
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
scikit-learn==1.1.2
scipy==1.9.0
seaborn==0.11.2
six==1.16.0
threadpoolctl==3.1.0
PyYAML==6.0